refactor: widen the last two accessors left computing in uint32
Context() and Sid() still recomputed their upper bound in uint32, against
the rule the rest of this branch follows.

Neither can wrap today: their guards bound the declared length by len(c), so
reaching 2^32 would take a >4 GiB buffer against a 16 MiB frame cap. This is
for the reader -- an accessor that looks like the bug next to twenty that no
longer do invites the wrong conclusion in review.

+++ internal/smb2/fscc.go
@@ -162,7 +162,7 @@ func (c SrvRequestResumeKeyResponseDecoder) ContextLength() uint32 {
 }
 
 func (c SrvRequestResumeKeyResponseDecoder) Context() []byte {
-	return c[28 : 28+c.ContextLength()]
+	return c[28 : 28+uint64(c.ContextLength())]
 }
 
 type SrvCopychunkCopy struct {
@@ -495,7 +495,7 @@ func (c FileQuotaInformationDecoder) QuotaLimit() int64 {
 }
 
 func (c FileQuotaInformationDecoder) Sid() SidDecoder {
-	return SidDecoder(c[40 : 40+c.SidLength()])
+	return SidDecoder(c[40 : 40+uint64(c.SidLength())])
 }
 
 type FileEndOfFileInformationEncoder struct {
